% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comarcas.R
\docType{data}
\name{comarcas}
\alias{comarcas}
\title{Ids das comarcas do TJPR}
\format{
tabela com 161 linhas e 2 colunas:
\describe{
\item{\code{id}}{id}
\item{\code{comarca}}{Nome da comarca}
}
}
\source{
\url{https://tjpr.jus.br/}
}
\usage{
comarcas
}
\description{
Ids das comarcas do TJPR
}
\keyword{datasets}
