% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjpr_baixar_dados_cjpg.R
\name{tjpr_baixar_dados_cjpg}
\alias{tjpr_baixar_dados_cjpg}
\title{Baixa tabela de sentencas do tjpr}
\usage{
tjpr_baixar_dados_cjpg(id_comarca, data_inicial, data_final, diretorio = ".")
}
\arguments{
\item{id_comarca}{Pode ser encontrada no dataset comarca}

\item{data_inicial}{Data inicial no formato dd/mm/yyyy}

\item{data_final}{Data final no formato dd/mm/yyyy}

\item{diretorio}{Diretório onde armazenar os htmls}
}
\value{
html
}
\description{
Baixa tabela de sentencas do tjpr
}
