% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjpr_extrair_dados_cjpg.R
\name{tjpr_extrair_dados_cjpg}
\alias{tjpr_extrair_dados_cjpg}
\title{Extrai as urls das sentenças(pdfs) do tjpr}
\usage{
tjpr_extrair_dados_cjpg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Arquivos htmls}

\item{diretorio}{Informar se arquivos forem NULL}
}
\value{
tibble
}
\description{
Extrai as urls das sentenças(pdfs) do tjpr
}
\details{
Esta função itera sobre todos os htmls informados
e busca na página do tjpr cada uma das urls dos pdfs.
Isso implica que ela pode demorar. Sugere-se não buscar
um número grande por vez.
}
